+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- [disk] fix crash due to insufficient permissions
+
 ## [0.5.1] - 2020-02-16
 
 ### Fixed